+++ config/global.php
@@ -56,7 +56,7 @@ return array(
         // If Piwik is not installed yet, it's possible the tmp/ folder is not writable
         // we prevent failing with an unclear message eg. coming from doctrine-cache
         // by forcing to use a cache backend which always works ie. array
-        if(!\Piwik\SettingsPiwik::isPiwikInstalled()) {
+        if(!\Piwik\SettingsPiwik::isMatomoInstalled()) {
             $backend = 'array';
         } else {
             try {